refactor(#1381): Move TestRandomNumberGenerator out into common testUtils
