+++ common/src/test/java/com/scottlogic/deg/testUtils/TestRandomNumberGenerator.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright 2019 Scott Logic Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.scottlogic.deg.testUtils;
+
+import com.scottlogic.deg.generator.utils.RandomNumberGenerator;
+
+import java.math.BigDecimal;
+
+public class TestRandomNumberGenerator implements RandomNumberGenerator {
+
+    private double nextDoubleValue = 0;
+
+    public void setNextDouble(double value) {
+        nextDoubleValue = value;
+    }
+
+    @Override
+    public int nextInt() {
+        return 0;
+    }
+
+    @Override
+    public int nextInt(int bound) {
+        return 0;
+    }
+
+    @Override
+    public int nextInt(int lowerInclusive, int upperExclusive) {
+        return 0;
+    }
+
+    @Override
+    public double nextDouble(double lower, double upper) {
+        return nextDoubleValue * (upper - lower) + lower;
+    }
+
+    @Override
+    public BigDecimal nextBigDecimal(BigDecimal lowerInclusive, BigDecimal upperExclusive) {
+        return new BigDecimal(nextDouble(lowerInclusive.doubleValue(), upperExclusive.doubleValue()));
+    }
+}
